Extract cache feed image data use case into the new `CacheFeedImageDataUseCaseTests` test case class
